LPS-117844 Update unit testes for EditTableView and ListTableView
