+++ modules/apps/app-builder/app-builder-web/test/js/constants.es.js
@@ -127,3 +127,87 @@ export const RESPONSES = {
 		totalCount: ITEMS.TWENTY.length + 1,
 	},
 };
+
+const dataDefinitionField = {
+	customProperties: {
+		autocomplete: false,
+		dataSourceType: 'manual',
+		dataType: 'string',
+		ddmDataProviderInstanceId: '[]',
+		ddmDataProviderInstanceOutput: '[]',
+		displayStyle: 'singleline',
+		fieldNamespace: '',
+		options: {
+			en_US: [
+				{
+					label: 'Option',
+					value: 'Option',
+				},
+			],
+		},
+		placeholder: {
+			en_US: '',
+		},
+		tooltip: {
+			en_US: '',
+		},
+		visibilityExpression: '',
+	},
+	defaultValue: {
+		en_US: '',
+	},
+	description: {
+		en_US: 'Enter your name',
+	},
+	fieldType: 'text',
+	indexType: 'keyword',
+	indexable: true,
+	label: {
+		en_US: 'Name',
+	},
+	localizable: true,
+	name: 'Text',
+	nestedDataDefinitionFields: [],
+	readOnly: false,
+	repeatable: false,
+	required: false,
+	showLabel: true,
+	tip: {
+		en_US: '',
+	},
+};
+
+const dataDefinition = {
+	availableLanguageIds: ['en_US'],
+	dataDefinitionFields: [dataDefinitionField],
+	dataDefinitionKey: '36601',
+	dateCreated: '2020-04-24T13:50:04Z',
+	dateModified: '2020-04-24T13:50:13Z',
+	defaultLanguageId: 'en_US',
+	description: {},
+	id: 36602,
+	name: {
+		en_US: 'My Custom Object',
+	},
+	siteId: 20125,
+	storageType: 'json',
+	userId: 20127,
+};
+
+export const DATA_DEFINITION_RESPONSES = {
+	ONE_ITEM: dataDefinition,
+	TWO_ITEMS: {
+		...dataDefinition,
+		dataDefinitionFields: [
+			dataDefinitionField,
+			{
+				...dataDefinitionField,
+				fieldType: 'select',
+				label: {
+					en_US: 'Options',
+				},
+				name: 'SelectFromList',
+			},
+		],
+	}
+};
