+++ modules/apps/app-builder/app-builder-web/test/js/pages/table-view/EditTableView.es.js
@@ -23,6 +23,7 @@ import {Route, Router} from 'react-router-dom';
 import EditTableView from '../../../../src/main/resources/META-INF/resources/js/pages/table-view/EditTableView.es';
 import * as toast from '../../../../src/main/resources/META-INF/resources/js/utils/toast.es';
 import AppContextProviderWrapper from '../../AppContextProviderWrapper.es';
+import {DATA_DEFINITION_RESPONSES} from '../../constants.es';
 
 const fieldTypes = [
 	{
@@ -84,87 +85,6 @@ const fieldTypeResponse = fieldTypes.map((fieldType, index) => ({
 	system: false,
 }));
 
-const dataDefinitionField = {
-	customProperties: {
-		autocomplete: false,
-		dataSourceType: 'manual',
-		dataType: 'string',
-		ddmDataProviderInstanceId: '[]',
-		ddmDataProviderInstanceOutput: '[]',
-		displayStyle: 'singleline',
-		fieldNamespace: '',
-		options: {
-			en_US: [
-				{
-					label: 'Option',
-					value: 'Option',
-				},
-			],
-		},
-		placeholder: {
-			en_US: '',
-		},
-		tooltip: {
-			en_US: '',
-		},
-		visibilityExpression: '',
-	},
-	defaultValue: {
-		en_US: '',
-	},
-	description: {
-		en_US: 'Enter your name',
-	},
-	fieldType: 'text',
-	indexType: 'keyword',
-	indexable: true,
-	label: {
-		en_US: 'Name',
-	},
-	localizable: true,
-	name: 'Text',
-	nestedDataDefinitionFields: [],
-	readOnly: false,
-	repeatable: false,
-	required: false,
-	showLabel: true,
-	tip: {
-		en_US: '',
-	},
-};
-
-const tableViewResponseOneItem = {
-	availableLanguageIds: ['en_US'],
-	dataDefinitionFields: [dataDefinitionField],
-	dataDefinitionKey: '36601',
-	dateCreated: '2020-04-24T13:50:04Z',
-	dateModified: '2020-04-24T13:50:13Z',
-	defaultLanguageId: 'en_US',
-	description: {},
-	id: 36602,
-	name: {
-		en_US: 'My Custom Object',
-	},
-	siteId: 20125,
-	storageType: 'json',
-	userId: 20127,
-};
-
-const tableViewResponseTwoItens = {
-	...tableViewResponseOneItem,
-	dataDefinitionFields: [
-		dataDefinitionField,
-		{
-			...dataDefinitionField,
-			fieldType: 'select',
-			label: {
-				en_US: 'Options',
-			},
-			name: 'SelectFromList',
-		},
-	],
-};
-
 const tableViewWithId = {
 	appliedFilters: {},
 	dataDefinitionId: 36716,
@@ -205,7 +125,7 @@ describe('EditTableView', () => {
 	it('renders', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseOneItem));
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.ONE_ITEM));
 
 		const {asFragment} = render(
 			<DndProvider backend={HTML5Backend}>
@@ -226,7 +146,7 @@ describe('EditTableView', () => {
 	it('renders with two fields in the sidebar and saves successfully', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseTwoItens))
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.TWO_ITEMS))
 			.mockResponseOnce();
 
 		const {queryAllByText, queryByPlaceholderText, queryByText} = render(
@@ -285,7 +205,7 @@ describe('EditTableView', () => {
 	it('renders with two fields in the sidebar and does not save successfully', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseTwoItens))
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.TWO_ITEMS))
 			.mockRejectOnce(() =>
 				Promise.reject(
 					JSON.stringify({
@@ -332,7 +252,7 @@ describe('EditTableView', () => {
 	it('renders with two fields in the sidebar and make actions', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseTwoItens));
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.TWO_ITEMS));
 
 		const {
 			container,
@@ -387,7 +307,7 @@ describe('EditTableView', () => {
 	it('renders with one field already inside the table and saves', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseTwoItens))
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.TWO_ITEMS))
 			.mockResponseOnce(JSON.stringify(tableViewWithId))
 			.mockResponseOnce(JSON.stringify({}));
 
@@ -435,7 +355,7 @@ describe('EditTableView', () => {
 	it('renders and toggle sidebar by clicking in the toggle button', async () => {
 		fetch
 			.mockResponseOnce(JSON.stringify(fieldTypeResponse))
-			.mockResponseOnce(JSON.stringify(tableViewResponseOneItem));
+			.mockResponseOnce(JSON.stringify(DATA_DEFINITION_RESPONSES.ONE_ITEM));
 
 		render(
 			<DndProvider backend={HTML5Backend}>
